moved table headers to one file
