+++ src/components/tableColumn/index.tsx
@@ -36,3 +36,49 @@ export const ReportColumns: ColumnsType = [
     }
   }
 ];
+
+export const MaturityTableCoulums: ColumnsType = [
+  {
+    title: 'Maturity Level',
+    dataIndex: 'maturityLevel'
+  },
+  {
+    title: 'Implementation Status',
+    dataIndex: 'implementationStatus'
+  },
+
+  {
+    title: 'Defensive Actions',
+    dataIndex: 'defensiveActions'
+  },
+
+  {
+    title: 'Offensive Actions',
+    dataIndex: 'offensiveActions'
+  },
+  {
+    title: 'Integrated Services',
+    dataIndex: 'integratedServices'
+  }
+];
+
+export const AppHardeningMaturityTableCoulums: ColumnsType = [
+  {
+    title: 'Maturity Level',
+    dataIndex: 'maturityLevel'
+  },
+  {
+    title: 'Implementation Status',
+    dataIndex: 'implementationStatus'
+  },
+
+  {
+    title: 'Defensive Actions',
+    dataIndex: 'defensiveActions'
+  },
+
+  {
+    title: 'Offensive Actions',
+    dataIndex: 'offensiveActions'
+  }
+];
